Update index.html
I added a function that creates an Image object, which acts as a palceholder for the image to generate in the canvas, and then added functions that input the image into the canvas when their respective button is clicked.
Also, I created my function for clearing or resetting the canvas; this took a lot of trial and error to get the dimensions of "ctx.clearRect(0,0, 600,300);" correct, but now the entire canvas is cleared when the button is pressed.
I also added two more buttons; one for the image and one to clear the canvas.